pyproject.toml
[tool.poetry]
name = "nonebot-plugin-bnhhsh"
version = "0.1.0"
description = "Se-se!"
authors = ["student_2333 <lgc2333@126.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
nonebot2 = ">=2.0.0b4"
nonebot-adapter-onebot = "^2.1.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
